pdf-extractor/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
av==14.1.0
blinker==1.9.0
branca==0.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.2.0
cryptography==44.0.2
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.115.11
ffmpeg-python==0.2.0
Flask==3.0.3
folium==0.19.4
fonttools==4.50.0
future==1.0.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.8.3
mysql-connector-python==8.3.0
narwhals==1.28.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.68.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.3.0
plotly==6.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyogrio==0.10.0
pyparsing==3.1.2
pypdfium2==4.30.1
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
requests==2.32.3
retrying==1.3.4
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.29.0
setuptools==75.8.0
shapely==2.0.7
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.46.1
tabulate==0.9.0
threadpoolctl==3.4.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
XlsxWriter==3.2.0
xyzservices==2025.1.0
zipp==3.21.0
